Adjust ratios for firstname frequency.
Added two new parts to the code to do this.

First I count all firstnames (not just a single initial) when the file is read into a map. Counts are corrected to account for different sized data than used by Torvik.

Then, when ratio is looked up, if firstname received the max score, then correct it as described by Torvik. r(x1) = r(x1)*n^(-1.0024)

Second part is that the load function fetch_records_from_txt has been overloaded to accept an argument as to whether or not firstnames need ot be adjusted for. If not, it calls the old fetch_records function.